tests/wpt/css-tests/css-shapes-1_dev/xhtml1/shape-outside-shape-inherit-000.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>Shape Outside Box Inherit Value</title>
        <link href="http://html.adobe.com/" rel="author" title="Adobe" />
        <link href="mailto:betravis@adobe.com" rel="author" title="Bear Travis" />
        <link href="mailto:stearns@adobe.com" rel="reviewer" title="Alan Stearns" /> <!-- 2014-03-04 -->
        <link href="http://www.w3.org/TR/css-shapes-1/#shape-outside-property" rel="help" />
        <meta content="Shape-outside takes can be assigned the 'inherit' value." name="assert" />
        <meta content="dom" name="flags" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="support/parsing-utils.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script type="text/javascript">
            test(function() {
                var outer = document.createElement("div");
                var inner = document.createElement("div");
                outer.appendChild(inner);
                outer.style.setProperty("shape-outside", "content-box");
                inner.style.setProperty("shape-outside", "inherit");
                document.body.appendChild(outer);
                var inline = inner.style.getPropertyValue("shape-outside");
                var style = getComputedStyle(inner);
                var computed = style.getPropertyValue("shape-outside");
                document.body.removeChild(outer);
                assert_equals(inline, "inherit");
                assert_equals(computed, "content-box");
            }, "shape-outside can be assigned 'inherit' value");
        </script>
    

</body></html>
